Separate the common parts of rest and SSH interfaces
